--- setup/website_event_sale_constraints/odoo/addons/website_event_sale_constraints
@@ -1,1 +1,0 @@
-../../../../website_event_sale_constraints
