+++ setup/website_event_sale_joined_constraint_single_attendee/odoo/addons/website_event_sale_joined_constraint_single_attendee
@@ -1,0 +1,1 @@
+../../../../website_event_sale_joined_constraint_single_attendee
