+++ setup/website_event_sale_joined_constraint_single_attendee/setup.py
