--- website_event_sale_constraints/controllers/__init__.py
@@ -1,1 +1,0 @@
-from . import main
