--- website_event_sale_constraints/models/__init__.py
@@ -1,3 +1,0 @@
-from . import event_event_ticket
-from . import product
-from . import website
