--- website_event_sale_constraints/readme/CONFIGURE.rst
@@ -1,3 +1,0 @@
-Go to a Product > Set product type = Event ticket > Tick "Is a child ticket" right above
-
-Child tickets can't be sold on website without a non-child ticket.
