--- website_event_sale_constraints/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-Allows to prevent website customer to buy specific tickets alone, without buying another type of tickets.
