--- website_event_sale_constraints/views/event_ticket_views.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!-- Copyright 2023 Camptocamp SA-->
-<!-- License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)-->
-<odoo>
-    <record id="event_event_ticket_view_tree_from_event" model="ir.ui.view">
-        <field name="model">event.event.ticket</field>
-        <field
-            name="inherit_id"
-            ref="event_sale.event_event_ticket_view_tree_from_event"
-        />
-        <field name="arch" type="xml">
-                <field name="price" position="after">
-                    <field name="is_child_ticket" />
-                </field>
-        </field>
-    </record>
-</odoo>
