--- website_event_sale_constraints/views/product_views.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2023 Camptocamp SA-->
-<!-- License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)-->
-<odoo>
-    <record id="product_template_form_view" model="ir.ui.view">
-        <field name="name">product.template.form</field>
-        <field name="model">product.template</field>
-        <field name="inherit_id" ref="product.product_template_form_view" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='detailed_type']" position="after">
-                <field
-                    name="is_child_ticket"
-                    string="Is a child ticket"
-                    attrs="{'invisible': [('detailed_type', '!=', 'event')]}"
-                />
-            </xpath>
-        </field>
-    </record>
-</odoo>
