+++ website_event_sale_joined_constraint_single_attendee/__init__.py
@@ -1,2 +1,1 @@
 from . import controllers
-from . import models
