+++ website_event_sale_joined_constraint_single_attendee/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import website_event_sale_joined_constraint_controller
